Fix l2x0 cache invalidate handling of unaligned addresses
Patch to fix l2x0 cache invalidate handling of unaligned addresses.
Applies to linux 2.6.22 for ARM-11 processors.
Available from ARM Linux Patch System:
  http://www.arm.linux.org.uk/developer/patches/
  ARM Patch ID 4568/1, Date: 14 Sep, 2007, Status: Applied.
Notes:
The l2x0_inv_range() function doesn't handle unaligned addresses
correctly. It's necessary to clean the cache lines that are at the
start and end of the invalidate range, if the addresses are not aligned,
to prevent corruption of other data sharing the same cache line.
Acked-by: Catalin Marinas

http://www.bitshrine.org/gpp/linux-2.6.22-mx-Fix-l2x0-cache-invalidate-handling-of-unal.patch