Bugzilla 677. Allow FEC MAC address override.
Patch for Bugzilla 677. Allow FEC MAC address override.
FEC MAC addresses on some MX27 TO2 chips are not preprogrammed
and cannot be overridden by Redboot.  This patch gets around
this issue by allowing the FEC MAC address to be specified on
the kernel command line using the following option:
  fec_mac=xx:xx:xx:xx:xx:xx
where the "xx"s are replaced with the hex MAC address listed
on the sticker for your board.  Valid MAC addresses from the
chip or from Redboot will not be overridden using this option.
Applies to linux 2.6.22 kernel on MX27 platforms.

http://www.bitshrine.org/gpp/linux-2.6.22-mx-Bugzilla-677.-Allow-FEC-MAC-address-overri.patch