ARM: vf610: Additional Power mangement fixes.
The standby sleep state puts the module in STOP mode which
allows to the SoC to be woken by a regular interrupt. To save
power the main PLL1 is bypassed and uses the 24MHz on-chip
oscillator. However, memory clock need to be at full speed,
hence use PLL2 only to keep memory clock. This provides the
best balance between low-power requirement while have the
ability to use any IRQ as wake-up source.

Move the pm_info struct to DDR RAM and don't try
to allocate SRAM.