ARM: dts: vf610: modify ocram0, ocram1 entries
MCC uses all of OCRAM0 (when loading an MQX image), and the first 64KB of
OCRAM1 (for shared data between the M4 and A5 core). The OCRAM device tree
definitions conflicted with MCC, such that MCC would fail to start when
attempting to request these memory regions.

This commit removes the OCRAM0 definition from the Vybrid dts, and removes the
first 64KB from the OCRAM1 definition, allowing MCC to function.

Signed-off-by: Anthony Felice <tony.felice@timesys.com>