mcc_vf610: isr: dequeue all available signals from signal queue
Whenever the ISR triggers, all signals from the signal queue should be
dequeued, then serviced, so that the signal queue is empty when the ISR
finishes.

This bug is an artifact of the mcc 1.x port. The mcc 1.x dequeue signal
function returned 1 on success. The mcc 2.x dequeue signal function returns
zero (MCC_SUCCESS) - but the isr operated on the mcc 1.x assumption that 1
was returned on success.

Signed-off-by: Anthony Felice <tony.felice@timesys.com>